Use tracex for all otel needs
This depends on https://github.com/snabble/go-utils/pull/91 landing first.

The goal is that getting the otel dependencies via tracex can help us conduct what version of all the otel libraries we use in our services. We've seen in the past that there can be trouble with conflicting versions.